prooflearn/ai-service/requirements.txt
flask
flask-cors
python-dotenv
groq
